Fixed a getConnection() bug caused by a silent autoboxing issue left over from when the long -> int client ID conversion was made.  Now connection lookups work again.
git-svn-id: https://jmonkeyengine.googlecode.com/svn/trunk@7152 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ engine/src/networking/com/jme3/network/Server.java
@@ -96,7 +96,7 @@ public interface Server
     /**
      *  Retrieves a hosted connection by ID.
      */
-    public HostedConnection getConnection( long id );     
+    public HostedConnection getConnection( int id );     
  
     /**
      *  Retrieves a read-only collection of all currently connected connections.
